+++ server/src/auth/auth.module.ts
@@ -6,8 +6,8 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { User } from "src/common/entities/user.entity";
 import { JwtService } from "@nestjs/jwt";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
-import { UserRepository } from "src/user/user.repository";
-import { AuthRepository } from "./auth.repository";
+import { UserRepository } from "src/common/repositories/user.repository";
+import { AuthRepository } from "../common/repositories/auth.repository";
 
 @Module({
     imports: [TypeOrmModule.forFeature([User]), TypeOrmCustomModule.forCustomRepository([UserRepository])],
