+++ server/src/common/types/course-data.ts
