+++ server/src/common/types/raw-recruit-data.ts
