+++ server/src/course/course.module.ts
@@ -4,7 +4,7 @@ import { CourseController } from "./course.controller";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
 import { Course } from "src/common/entities/course.entity";
-import { CourseRepository } from "./course.repository";
+import { CourseRepository } from "../common/repositories/course.repository";
 
 @Module({
     imports: [TypeOrmModule.forFeature([Course]), TypeOrmCustomModule.forCustomRepository([CourseRepository])],
