+++ server/src/course/dto/create-course.dto.ts
@@ -1,6 +1,6 @@
 import { IsNumber, IsNumberString, IsString } from "class-validator";
-import { isValidPath } from "src/common/decorator/path.validator";
-import { LatLng } from "src/common/type/lat-lng";
+import { isValidPath } from "src/common/decorators/path.validator";
+import { LatLng } from "src/common/types/lat-lng";
 import { Course } from "src/common/entities/course.entity";
 
 export class CreateCourseDto {
