+++ server/src/recruit/dto/create-recruit.dto.ts
@@ -1,6 +1,6 @@
 import { Type } from "class-transformer";
 import { IsNumber, IsNumberString, IsString } from "class-validator";
-import { IsValidDateTime } from "src/common/decorator";
+import { IsValidDateTime } from "src/common/decorators";
 import { Recruit } from "../../common/entities/recruit.entity";
 
 export class CreateRecruitDto {
