+++ server/src/recruit/recruit.module.ts
@@ -4,10 +4,10 @@ import { RecruitController } from "./recruit.controller";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
 import { AuthService } from "src/auth/auth.service";
 import { JwtService } from "@nestjs/jwt";
-import { AuthRepository } from "src/auth/auth.repository";
-import { UserRepository } from "src/user/user.repository";
+import { AuthRepository } from "src/common/repositories/auth.repository";
+import { UserRepository } from "src/common/repositories/user.repository";
 import { UserRecruitRepository } from "src/user_recruit.repository";
-import { RecruitRepository } from "./recruit.repository";
+import { RecruitRepository } from "../common/repositories/recruit.repository";
 
 @Module({
     imports: [TypeOrmCustomModule.forCustomRepository([RecruitRepository, UserRepository, UserRecruitRepository])],
