+++ server/src/user/dto/create-user.dto.ts
@@ -1,6 +1,6 @@
 import { Type } from "class-transformer";
 import { IsNumber, IsNumberString } from "class-validator";
-import { IsValidId, IsValidPassword } from "src/common/decorator";
+import { IsValidId, IsValidPassword } from "src/common/decorators";
 import { User } from "../../common/entities/user.entity";
 
 export class CreateUserDto {
