+++ server/src/user/user.module.ts
@@ -4,7 +4,7 @@ import { UserController } from "./user.controller";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
 import { User } from "src/common/entities/user.entity";
-import { UserRepository } from "./user.repository";
+import { UserRepository } from "../common/repositories/user.repository";
 import { UserRecruitRepository } from "src/user_recruit.repository";
 
 @Module({
